"Disabled error squiggles in C++ and updated settings format"
This commit contains a change to the settings of a Visual Studio Code workspace. The 'files.associations' section has been updated with an additional comma to allow for the addition of a new setting on the next line. The new setting added is 'C\_Cpp.errorSquiggles' and it has been set to 'disabled'. This means that the error squiggles (visual cues for syntax or other errors) will no longer be shown for C++ files in this workspace.